fix: Add missing args into stories
